% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{blueAvatar}
\alias{blueAvatar}
\title{Create a Boostrap 4 avatar}
\usage{
blueAvatar(src, toolTip = NULL, size = c("sm", "md", "lg"))
}
\arguments{
\item{src}{Avatar inag path or url.}

\item{toolTip}{Optional tooltip triggered on mouse hover.}

\item{size}{Avatar size: choose one of "sm", "md" or "lg".}
}
\description{
Build an blue avatar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
