% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{blueAvatars}
\alias{blueAvatars}
\title{Create a Boostrap 4 avatar container}
\usage{
blueAvatars(...)
}
\arguments{
\item{...}{Slot for \link{blueAvatar}.}
}
\description{
Build an blue avatar container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
