% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{blueDashBody}
\alias{blueDashBody}
\title{Create a Boostrap 4 dashboard body}
\usage{
blueDashBody(...)
}
\arguments{
\item{...}{Body content, slot for \link{blueTabItems}}
}
\description{
Build an blue dashboard body
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
