% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{blueDashFooter}
\alias{blueDashFooter}
\title{Create a Boostrap 4 dashboard footer}
\usage{
blueDashFooter(..., copyrights = NULL, src = NULL)
}
\arguments{
\item{...}{Slot for \link{blueFooterMenu}.}

\item{copyrights}{Left text, if any.}

\item{src}{Left text src, if any.}
}
\description{
Build an blue dashboard footer
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
