% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueDashGallery.R
\name{blueDashGallery}
\alias{blueDashGallery}
\title{Launch the blueDash Gallery}
\usage{
blueDashGallery()
}
\description{
A gallery of all components available in blueDash.
}
\examples{

if (interactive()) {

 blueDashGallery()

}
}
