% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{blueDashNavbar}
\alias{blueDashNavbar}
\title{Create a Boostrap 4 dashboard navbar}
\usage{
blueDashNavbar(...)
}
\arguments{
\item{...}{Any element such as \link{blueDropNav}.}
}
\description{
Build an blue dashboard navbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
