% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{blueDropNav}
\alias{blueDropNav}
\title{Create a Boostrap 4 dashboard dropdown menu}
\usage{
blueDropNav(..., title = NULL, src = NULL, orientation = "right")
}
\arguments{
\item{...}{slot for \link{blueDropNavItem}, \link{blueDropNavTitle} or \link{blueDropNavDivider}.}

\item{title}{Menu title.}

\item{src}{Menu image, if any.}

\item{orientation}{Dropdown orientation: "left" or "right".}
}
\description{
Build an blue dashboard dropdown menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
