% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{blueDropNavDivider}
\alias{blueDropNavDivider}
\title{Create a Boostrap 4 dashboard dropdown menu divider}
\usage{
blueDropNavDivider()
}
\description{
Build an blue dashboard dropdown menu divider
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
