% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{blueDropNavItem}
\alias{blueDropNavItem}
\title{Create a Boostrap 4 dashboard dropdown menu item}
\usage{
blueDropNavItem(title = NULL, src = NULL, icon = NULL)
}
\arguments{
\item{title}{Item title.}

\item{src}{External link.}

\item{icon}{Icon, if any. Expect \link{blueIcon} or \link[shiny]{icon}.}
}
\description{
Build an blue dashboard dropdown menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
