% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{blueDropNavTitle}
\alias{blueDropNavTitle}
\title{Create a Boostrap 4 dashboard dropdown menu title}
\usage{
blueDropNavTitle(title = NULL)
}
\arguments{
\item{title}{Menu title.}
}
\description{
Build an blue dashboard dropdown menu title
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
