% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{blueFooterItem}
\alias{blueFooterItem}
\title{Create a Boostrap 4 dashboard footer menu item}
\usage{
blueFooterItem(..., src = NULL)
}
\arguments{
\item{...}{Item name.}

\item{src}{Item external link.}
}
\description{
Build an blue dashboard footer menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
