% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{blueFooterMenu}
\alias{blueFooterMenu}
\title{Create a Boostrap 4 dashboard footer menu}
\usage{
blueFooterMenu(...)
}
\arguments{
\item{...}{Slot for \link{blueFooterItem}.}
}
\description{
Build an blue dashboard footer menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
