% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{blueSidebarHeader}
\alias{blueSidebarHeader}
\title{A header to put inside \link{blueDashSidebar}}
\usage{
blueSidebarHeader(title = NULL)
}
\arguments{
\item{title}{Header title}
}
\description{
A header to put inside \link{blueDashSidebar}
}
