% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{blueSidebarItem}
\alias{blueSidebarItem}
\title{Create a Boostrap 4 blue sidebar menu}
\usage{
blueSidebarItem(..., tabName = NULL, icon = NULL)
}
\arguments{
\item{...}{Item name.}

\item{tabName}{Should correspond exactly to the tabName given in \code{\link{blueTabItem}}.}

\item{icon}{Item icon. \link{blueIcon} or \link[shiny]{icon}.}
}
\description{
Build an blue dashboard main sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
