% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{blueSidebarMenu}
\alias{blueSidebarMenu}
\title{Create a Boostrap 4 blue sidebar menu}
\usage{
blueSidebarMenu(...)
}
\arguments{
\item{...}{Slot for \link{blueSidebarItem}.}
}
\description{
Build an blue dashboard main sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
