% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{blueTabItem}
\alias{blueTabItem}
\title{One tab to put inside a tab items container}
\usage{
blueTabItem(tabName = NULL, ...)
}
\arguments{
\item{tabName}{The name of a tab. This must correspond to the \code{tabName}
of a \code{\link{blueSidebarItem}}.}

\item{...}{Contents of the tab.}
}
\description{
One tab to put inside a tab items container
}
