% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{blueTabItems}
\alias{blueTabItems}
\title{A container for tab items}
\usage{
blueTabItems(...)
}
\arguments{
\item{...}{Items to put in the container. Each item should be a
\code{\link{blueTabItem}}.}
}
\description{
A container for tab items
}
