% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{blueTable}
\alias{blueTable}
\title{Create a Boostrap 4 table container}
\usage{
blueTable(
  ...,
  cardWrap = FALSE,
  title = NULL,
  headTitles,
  dark = FALSE,
  width = 12
)
}
\arguments{
\item{...}{\link{blueTableItems}.}

\item{cardWrap}{Whether to wrap the table in a card. FALSE by default.}

\item{title}{Card title if any.}

\item{headTitles}{Table header names. Must have the same length as the number of 
\link{blueTableItem} in \link{blueTableItems}. Set "" to have an empty title field.}

\item{dark}{Whether to enable dark mode. FALSE by default.}

\item{width}{Table width. 12 by default.}
}
\description{
Build an blue table container
}
\examples{
if (interactive()) {
 library(shiny)
 library(blueR)
 library(blueDash)
 shinyApp(
  ui = blueDashPage(
    navbar = blueDashNavbar(), 
    sidebar = blueDashSidebar(id = "mysidebar"), 
    header = blueDashHeader(), 
    body = blueDashBody(
     blueTable(
      headTitles = c(
       "PROJECT",
       "BUDGET",
       "STATUS",
       "USERS",
       "COMPLETION",
       ""
      ),
      blueTableItems(
       blueTableItem("Blue Design System"),
       blueTableItem(dataCell = TRUE, "$2,500 USD"),
       blueTableItem(
        dataCell = TRUE, 
        blueBadge(
         text = "Pending",
         status = "danger"
        )
       ),
       blueTableItem(
        blueAvatar(
         size = "sm",
         src = "https://image.flaticon.com/icons/svg/219/219976.svg"
         )
       ),
       blueTableItem(
        dataCell = TRUE, 
        blueProgress(value = 60, status = "danger")
       ),
       blueTableItem(
        blueButton(
         name = "Click me!",
         status = "warning",
         icon = "atom",
         size = "sm"
        )
       )
      )
     )
    ), 
    footer = blueDashFooter()
  ),
  server = function(input, output) { }
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
