% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{blueTableItem}
\alias{blueTableItem}
\title{Create a Boostrap 4 table item}
\usage{
blueTableItem(..., dataCell = FALSE)
}
\arguments{
\item{...}{Any HTML element.}

\item{dataCell}{Whether the cell should be contain data or text. <td> by default.}
}
\description{
Build an blue table item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
