% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{blueTableItems}
\alias{blueTableItems}
\title{Create a Boostrap 4 table item row}
\usage{
blueTableItems(...)
}
\arguments{
\item{...}{Slot for \link{blueTableItem}.}
}
\description{
Build an blue table item row
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
